project v0.9

+++ list.php
@@ -34,25 +34,28 @@ $rows = $result->fetchAll();
 
 <div class="admin-menu">
             <div>
-                <a class='btn btn-sm filter-button' href="list.php?list=wallpapers">Tapety</a>
-                <a class='btn btn-sm filter-button' href="list.php?list=categories">Kategorie</a>
-                <a class='btn btn-sm filter-button' href="list.php?list=users">Użytkownicy</a>
+                <a class= '<?php if ($list == 'wallpapers') {echo "active-class ";} ?>btn btn-sm filter-button' href="list.php?list=wallpapers">Tapety</a>
+                <a class='<?php if ($list == 'categories') {echo "active-class ";} ?>btn btn-sm filter-button' href="list.php?list=categories">Kategorie</a>
+                <a class='<?php if ($list == 'users') {echo "active-class ";} ?>btn btn-sm filter-button' href="list.php?list=users">Użytkownicy</a>
 </div>
         </div>
         <?php 
         if ($list === 'wallpapers') {
+            echo "<h1>Tapety</h1><br>";
             echo "<a class='btn btn-primary add-button' href='admin.php?list=$list'>Dodaj nową tapetę</a>";
             } 
         
         else if ($list === 'categories') {
+            echo "<h1>Kategorie</h1><br>";
             echo "<a class='btn btn-primary add-button' href='admin.php?list=$list'>Dodaj nową kategorię</a>";    
             } 
             else if ($list === 'users') {
+            echo "<h1>Użytkownicy</h1><br>";
             echo "<a class='btn btn-primary add-button' href='admin.php?list=$list'>Dodaj nowego użytkownika</a>";    
             }?>
         <div class='container'>
             <div class="row cold-md-offset-2 custyle justify-content-center col-auto">
-<!-- <table class="styled-table"> -->
+
     <table class="table table-responsive custab styled-table">
     <thead>
         <tr>
@@ -72,11 +75,11 @@ $rows = $result->fetchAll();
             echo "<tr>";
             echo "<td>$id</td>";
             echo "<td>$item_date</td>";
-            echo "<td>$name</td>";
+            echo "<td><a href='wallpaper.php?id=$id'>$name</a></td>";
             echo "<td>
                 <div class='btn-group text-lg-start'>
                     <a class='btn btn-sm btn-secondary action-button' style='width: fit-content' href='edit.php?id=$id&list=$list'>Edytuj</a>
-                    <a class='btn btn-sm btn-danger action-button' style='width:fit-content' href='delete.php?id=$id&list=$list'>Usuń</a>
+                    <a class='btn btn-sm btn-danger action-button delete-button' data-confirm='Czy na pewno chcesz usunąć ten element?' style='width:fit-content' href='delete.php?id=$id&list=$list'>Usuń</a>
                     </div>
                     </td>";
             echo "</tr>";
